% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{coef.lookup}
\alias{coef.lookup}
\title{Coefficient lookup.}
\usage{
\method{coef}{lookup}(variable, value)
}
\arguments{
\item{variable}{A column name from the original data frame.  This will be the
variable one of whose values you are looking up the coefficient for.}

\item{value}{A specific value of the variable under consideration.  This value
must be one of the values actually used in the data frame for 
\code{fit.model}.}
}
\value{
A numeric value, zero for a baseline category or the estimated
 coefficient from the fitted logistic model.  These estimates are logarithms
 of the cumulative logit (once you add the relevent cut).
}
\description{
This utility function is used by other functions to lookup coefficients that
have been estimated by \code{fit.model}.  The estimates are found in the 
\code{polr} class object \code{fit}.  Baseline categories return a value of
zero.
}
\details{
The function checks that the variable is in the fitted model, and that the
value for that variable is one that has been included in the model.
}
\examples{
coef.lookup("gender","female")
}

